[level 1] Title: 3진법 뒤집기, Time: 0.07 ms, Memory: 16.4 MB -BaekjoonHub

+++ 프로그래머스/1/68935. 3진법 뒤집기/3진법 뒤집기.swift
@@ -30,7 +30,7 @@ func solution(_ n:Int) -> Int {
             if cnt == 0 {
                 result += res
             } else {
-                result += Int(Float(res) * pow(Float(3), Float(cnt)))
+                result += Int(Double(res) * pow(Double(3), Double(cnt)))
             }
         }
         cnt += 1
